requirements.txt
requests
openai
google-generativeai
faiss-cpu
sentence-transformers
streamlit
